requirements.txt
Flask==2.0.3
tensorflow-hub
tensorflow
google-cloud-storage
pillow
numpy
werkzeug==2.0.3
